.gitignore: broaden build/ to build*/ to allow several build directories
In daily development work with CMake it may be convenient to create several build
directories. That's not an uncommon practice to have, for example, 'build' for
ordinary build, 'build-debug' for debug build, 'build-qtcreator-debug' when using
an IDE, or 'build-arm' specifically for cross-compilation. Hence, 'build*/' would
be a handier .gitignore pattern.

Signed-off-by: Pavel Kretov <firegurafiku@gmail.com>